+++ lib/routes/hwyd_routes.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/cupertino.dart';
-import 'package:hwyd/page/journal_page.dart';
+import 'package:hwyd/screens/journal_page.dart';
 
 Map<String, WidgetBuilder> routes = <String, WidgetBuilder>{
   '/journal': (BuildContext context) => JournalPage(),
