+++ test/hwyd_smoke_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:hwyd/page/journal_page.dart';
+import 'package:hwyd/screens/journal_page.dart';
 
 void main() {
   testWidgets('Hwyd smoke test', (WidgetTester tester) async {
